// types/games.ts
export type games = {
	id: number;
	name: string;
	genre: string;
	photoGame: string;
	headerImage: string;
	description: string;
	platform: string;
	producer: string;
	studio: string;
	language: string;
	status: string;
	price: number;
	rating: number;
	popularity: number;
	releaseDate: string;
	istrending: boolean;
	shopLink: string;
	trailerUrl: string;
	studiophoto: string;
	storyline: string;
};